+++ .travis.yml
@@ -7,8 +7,8 @@ notifications:
 script:
 - set -o pipefail
 - cd Support
-- xcodebuild -project AppInsights.xcodeproj -scheme "AppInsights" -sdk iphonesimulator
+- xcodebuild -project ApplicationInsights.xcodeproj -scheme "ApplicationInsights" -sdk iphonesimulator
   -configuration Debug -destination "platform=iOS Simulator,name=iPhone 6" test |
   xcpretty -c
-- xcodebuild -project AppInsights.xcodeproj -scheme "AppInsights Framework" -sdk iphonesimulator
+- xcodebuild -project ApplicationInsights.xcodeproj -scheme "ApplicationInsights Framework" -sdk iphonesimulator
   -configuration Release | xcpretty -c
