+++ ApplicationInsights.podspec
@@ -7,10 +7,10 @@ Pod::Spec.new do |s|
                        This SDK will allow you to send telemetry of various kinds (event, trace, exception, etc.) and useful crash reports to the Application Insights service where they can be visualized in the Azure Portal.
                        DESC
   s.homepage         = "https://github.com/Microsoft/AppInsights-iOS/"
-  s.license          = { :type => 'MIT', :file => 'AppInsights/LICENSE' }
+  s.license          = { :type => 'MIT', :file => 'ApplicationInsights/LICENSE' }
   s.author           = { "Microsoft" => "appinsights-ios@microsoft.com" }
 
-  s.source           = { :http => "https://github.com/Microsoft/AppInsights-iOS/releases/download/v#{s.version}/AppInsights-#{s.version}.zip" }
+  s.source           = { :http => "https://github.com/Microsoft/AppInsights-iOS/releases/download/v#{s.version}/ApplicationInsights-#{s.version}.zip" }
 
   s.platform        = :ios, '6.0'
   s.requires_arc    = true
@@ -18,6 +18,6 @@ Pod::Spec.new do |s|
   s.frameworks      = 'UIKit', 'Foundation', 'SystemConfiguration', 'Security'
   s.weak_framework  = 'CoreTelephony'
 
-  s.ios.vendored_frameworks = 'AppInsights/AppInsights.framework'
-  s.preserve_path   = 'AppInsights/README.md'
+  s.ios.vendored_frameworks = 'ApplicationInsights/ApplicationInsights.framework'
+  s.preserve_path   = 'ApplicationInsights/README.md'
 end
