+++ Classes/ApplicationInsightsFeatureConfig.h
