+++ Classes/ApplicationInsightsPrivate.h
@@ -1,5 +1,5 @@
 #import <Foundation/Foundation.h>
-#import "MSAIAppInsights.h"
+#import "MSAIApplicationInsights.h"
 
 #ifndef MSAI_Private_h
 #define MSAI_Private_h
@@ -21,7 +21,7 @@ extern NSString *const kMSAIIntegrationflowTimestamp;
 
 extern NSString *const kMSAITelemetryPath;
 
-#define MSAILog(fmt, ...) do { if([MSAIAppInsights sharedInstance].isDebugLogEnabled && ![MSAIAppInsights sharedInstance].isAppStoreEnvironment) { NSLog((@"[MSAI] %s/%d " fmt), __PRETTY_FUNCTION__, __LINE__, ##__VA_ARGS__); }} while(0)
+#define MSAILog(fmt, ...) do { if([MSAIApplicationInsights sharedInstance].isDebugLogEnabled && ![MSAIApplicationInsights sharedInstance].isAppStoreEnvironment) { NSLog((@"[MSAI] %s/%d " fmt), __PRETTY_FUNCTION__, __LINE__, ##__VA_ARGS__); }} while(0)
 
 #ifndef __IPHONE_8_0
 #define __IPHONE_8_0     80000
