+++ Classes/MSAICrashDataProvider.h
@@ -12,7 +12,7 @@
 
 
 /**
- *  AppInsights Crash Reporter error domain
+ *  ApplicationInsights Crash Reporter error domain
  */
 typedef NS_ENUM (NSInteger, MSAIBinaryImageType) {
   /**
