+++ Classes/MSAICrashDetails.h
@@ -69,7 +69,7 @@
  
  See `[MSAICrashManager appNotTerminatingCleanlyDetectionEnabled]` for more details about which kind of kills can be detected.
  
- @warning This property only has a correct value, once `[MSAIAppInsights start]` was
+ @warning This property only has a correct value, once `[MSAIApplicationInsights start]` was
  invoked! In addition, it is automatically disabled while a debugger session is active!
  
  @see `[MSAICrashManager appNotTerminatingCleanlyDetectionEnabled]`
