+++ Classes/MSAICrashManager.m
@@ -1,8 +1,8 @@
-#import "AppInsights.h"
+#import "ApplicationInsights.h"
 
 #if MSAI_FEATURE_CRASH_REPORTER
 
-#import "AppInsightsPrivate.h"
+#import "ApplicationInsightsPrivate.h"
 #import "MSAIHelper.h"
 #import "MSAICrashManagerPrivate.h"
 #import "MSAICrashDataProvider.h"
@@ -113,7 +113,7 @@ static PLCrashReporterCallbacks defaultCallback = {
 
       if(!msai_isAppStoreEnvironment()) {
         if(self.debuggerIsAttached) {
-          NSLog(@"[AppInsights] WARNING: Detecting crashes is NOT enabled due to running the app with a debugger attached.");
+          NSLog(@"[ApplicationInsights] WARNING: Detecting crashes is NOT enabled due to running the app with a debugger attached.");
         }
       }
 
@@ -181,7 +181,7 @@ static PLCrashReporterCallbacks defaultCallback = {
     // and can show a debug warning log message, that the dev has to make sure the "newer" error handler
     // doesn't exit the process itself, because then all subsequent handlers would never be invoked.
     //
-    // Note: ANY error handler setup BEFORE AppInsights initialization will not be processed!
+    // Note: ANY error handler setup BEFORE ApplicationInsights initialization will not be processed!
 
     // get the current top level error handler
     NSUncaughtExceptionHandler *initialHandler = NSGetUncaughtExceptionHandler();
@@ -199,7 +199,7 @@ static PLCrashReporterCallbacks defaultCallback = {
 
     // Enable the Crash Reporter
     if(![self.plCrashReporter enableCrashReporterAndReturnError:&error]) {
-      NSLog(@"[AppInsights] WARNING: Could not enable crash reporter: %@", [error localizedDescription]);
+      NSLog(@"[ApplicationInsights] WARNING: Could not enable crash reporter: %@", [error localizedDescription]);
     }
 
     // get the new current top level error handler, which should now be the one from PLCrashReporter
@@ -212,7 +212,7 @@ static PLCrashReporterCallbacks defaultCallback = {
       MSAILog(@"INFO: Exception handler successfully initialized.");
     } else {
       // this should never happen, theoretically only if NSSetUncaugtExceptionHandler() has some internal issues
-      NSLog(@"[AppInsights] ERROR: Exception handler could not be set. Make sure there is no other exception handler set up!");
+      NSLog(@"[ApplicationInsights] ERROR: Exception handler could not be set. Make sure there is no other exception handler set up!");
     }
   }
 }
@@ -315,7 +315,7 @@ void msai_save_events_callback(siginfo_t *info, ucontext_t *uap, void *context) 
     name[3] = getpid();
 
     if(sysctl(name, 4, &info, &info_size, NULL, 0) == -1) {
-      NSLog(@"[AppInsights] ERROR: Checking for a running debugger via sysctl() failed: %s", strerror(errno));
+      NSLog(@"[ApplicationInsights] ERROR: Checking for a running debugger via sysctl() failed: %s", strerror(errno));
       debuggerIsAttached = false;
     }
 
@@ -330,7 +330,7 @@ void msai_save_events_callback(siginfo_t *info, ucontext_t *uap, void *context) 
   if(!msai_isAppStoreEnvironment()) {
 
     if(self.debuggerIsAttached) {
-      NSLog(@"[AppInsights] WARNING: The debugger is attached. The following crash cannot be detected by the SDK!");
+      NSLog(@"[ApplicationInsights] WARNING: The debugger is attached. The following crash cannot be detected by the SDK!");
     }
 
     __builtin_trap();
@@ -528,7 +528,7 @@ void msai_save_events_callback(siginfo_t *info, ucontext_t *uap, void *context) 
     // If the top level error handler differs from our own, then at least another one was added.
     // This could cause exception crashes not to be reported to HockeyApp. See log message for details.
     if (self.exceptionHandler != currentHandler) {
-      NSLog(@"[AppInsights] ERROR: Exception handler could not be set. Make sure there is no other exception handler set up!");
+      NSLog(@"[ApplicationInsights] ERROR: Exception handler could not be set. Make sure there is no other exception handler set up!");
     }
   }
 }
