+++ Support/AppInsightsTests/ApplicationInsightsTests-Info.plist
