+++ Support/AppInsightsTests/MSAIHelperTests.m
@@ -6,10 +6,10 @@
 #define MOCKITO_SHORTHAND
 #import <OCMockitoIOS/OCMockitoIOS.h>
 
-#import "AppInsights.h"
+#import "ApplicationInsights.h"
 #import "MSAIHelper.h"
 #import "MSAIKeychainUtils.h"
-#import "AppInsightsPrivate.h"
+#import "ApplicationInsightsPrivate.h"
 
 
 @interface MSAIHelperTests : XCTestCase
