+++ Support/AppInsightsTests/MSAIKeychainUtilsTests.m
@@ -6,7 +6,7 @@
 #define MOCKITO_SHORTHAND
 #import <OCMockitoIOS/OCMockitoIOS.h>
 
-#import "AppInsights.h"
+#import "ApplicationInsights.h"
 #import "MSAIKeychainUtils.h"
 
 @interface MSAIKeychainUtilsTests : XCTestCase {
