+++ Support/AppInsightsTests/MSAINotificationTests.h
@@ -6,7 +6,7 @@
 #define MOCKITO_SHORTHAND
 #import <OCMockitoIOS/OCMockitoIOS.h>
 
-@interface NotificationTests : XCTestCase
+@interface MSAINotificationTests : XCTestCase
 
 - (void)setMockNotificationCenter:(NSNotificationCenter *)notificationCenter;
 - (NSNotificationCenter *)mockNotificationCenter;
