+++ Support/AppInsightsTests/MSAIPersistenceTest.m
@@ -6,7 +6,7 @@
 #define MOCKITO_SHORTHAND
 #import <OCMockitoIOS/OCMockitoIOS.h>
 
-#import "AppInsights.h"
+#import "ApplicationInsights.h"
 #import "MSAIEnvelope.h"
 #import "MSAIPersistence.h"
