+++ Support/ApplicationInsights.xcconfig
