+++ Support/ApplicationInsights.xcodeproj/project.xcworkspace/contents.xcworkspacedata
@@ -2,6 +2,6 @@
 <Workspace
    version = "1.0">
    <FileRef
-      location = "self:AppInsights.xcodeproj">
+      location = "self:ApplicationInsights.xcodeproj">
    </FileRef>
 </Workspace>
