+++ Support/ApplicationInsights.xcodeproj/project.xcworkspace/xcshareddata/AppInsights-iOS.xccheckout
