+++ Support/ApplicationInsights.xcodeproj/project.xcworkspace/xcshareddata/AppInsights.xccheckout
