+++ Support/ApplicationInsights.xcodeproj/xcshareddata/xcbaselines/1E5A458F16F0DFC200B55C04.xcbaseline/0E03ADBE-1442-48B6-ADD5-807E0C73245B.plist
