+++ Support/ApplicationInsights.xcodeproj/xcshareddata/xcbaselines/1E5A458F16F0DFC200B55C04.xcbaseline/5C0AC091-0676-4E63-8E6F-B855856B85CA.plist
