+++ Support/ApplicationInsights.xcodeproj/xcshareddata/xcschemes/ApplicationInsights Tests.xcscheme
@@ -1,6 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <Scheme
-   LastUpgradeVersion = "0610"
    version = "1.3">
    <BuildAction
       parallelizeBuildables = "YES"
@@ -9,15 +8,15 @@
          <BuildActionEntry
             buildForTesting = "YES"
             buildForRunning = "YES"
-            buildForProfiling = "NO"
-            buildForArchiving = "NO"
+            buildForProfiling = "YES"
+            buildForArchiving = "YES"
             buildForAnalyzing = "YES">
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "1E5A458F16F0DFC200B55C04"
-               BuildableName = "AppInsightsTests.xctest"
-               BlueprintName = "AppInsights Tests"
-               ReferencedContainer = "container:AppInsights.xcodeproj">
+               BuildableName = "ApplicationInsightsTests.xctest"
+               BlueprintName = "ApplicationInsights Tests"
+               ReferencedContainer = "container:ApplicationInsights.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -33,21 +32,12 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "1E5A458F16F0DFC200B55C04"
-               BuildableName = "AppInsightsTests.xctest"
-               BlueprintName = "AppInsights Tests"
-               ReferencedContainer = "container:AppInsights.xcodeproj">
+               BuildableName = "ApplicationInsightsTests.xctest"
+               BlueprintName = "ApplicationInsights Tests"
+               ReferencedContainer = "container:ApplicationInsights.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
-      <MacroExpansion>
-         <BuildableReference
-            BuildableIdentifier = "primary"
-            BlueprintIdentifier = "1E5A458F16F0DFC200B55C04"
-            BuildableName = "AppInsightsTests.xctest"
-            BlueprintName = "AppInsights Tests"
-            ReferencedContainer = "container:AppInsights.xcodeproj">
-         </BuildableReference>
-      </MacroExpansion>
    </TestAction>
    <LaunchAction
       selectedDebuggerIdentifier = "Xcode.DebuggerFoundation.Debugger.LLDB"
@@ -62,9 +52,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "1E5A458F16F0DFC200B55C04"
-            BuildableName = "AppInsightsTests.xctest"
-            BlueprintName = "AppInsights Tests"
-            ReferencedContainer = "container:AppInsights.xcodeproj">
+            BuildableName = "ApplicationInsightsTests.xctest"
+            BlueprintName = "ApplicationInsights Tests"
+            ReferencedContainer = "container:ApplicationInsights.xcodeproj">
          </BuildableReference>
       </MacroExpansion>
       <AdditionalOptions>
@@ -76,15 +66,6 @@
       useCustomWorkingDirectory = "NO"
       buildConfiguration = "Release"
       debugDocumentVersioning = "YES">
-      <MacroExpansion>
-         <BuildableReference
-            BuildableIdentifier = "primary"
-            BlueprintIdentifier = "1E5A458F16F0DFC200B55C04"
-            BuildableName = "AppInsightsTests.xctest"
-            BlueprintName = "AppInsights Tests"
-            ReferencedContainer = "container:AppInsights.xcodeproj">
-         </BuildableReference>
-      </MacroExpansion>
    </ProfileAction>
    <AnalyzeAction
       buildConfiguration = "Debug">
