+++ Support/ApplicationInsights.xcodeproj/xcshareddata/xcschemes/ApplicationInsights.xcscheme
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <Scheme
-   version = "1.7">
+   version = "1.3">
    <BuildAction
       parallelizeBuildables = "YES"
       buildImplicitDependencies = "YES">
@@ -14,9 +14,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "1E5954CB15B6F24A00A03429"
-               BuildableName = "libAppInsights.a"
-               BlueprintName = "AppInsights"
-               ReferencedContainer = "container:AppInsights.xcodeproj">
+               BuildableName = "libApplicationInsights.a"
+               BlueprintName = "ApplicationInsights"
+               ReferencedContainer = "container:ApplicationInsights.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -25,35 +25,16 @@
       selectedDebuggerIdentifier = "Xcode.DebuggerFoundation.Debugger.LLDB"
       selectedLauncherIdentifier = "Xcode.DebuggerFoundation.Launcher.LLDB"
       shouldUseLaunchSchemeArgsEnv = "YES"
-      buildConfiguration = "CodeCoverage">
-      <PostActions>
-         <ExecutionAction
-            ActionType = "Xcode.IDEStandardExecutionActionsCore.ExecutionActionType.ShellScriptAction">
-            <ActionContent
-               title = "Run Script"
-               scriptText = "if [ &quot;${CONFIGURATION}&quot; = &quot;CodeCoverage&quot; ]; then&#10;    source ${SRCROOT}/../Vendor/XcodeCoverage/run_code_coverage_post.sh&#10;fi"
-               shellToInvoke = "/bin/bash">
-               <EnvironmentBuildable>
-                  <BuildableReference
-                     BuildableIdentifier = "primary"
-                     BlueprintIdentifier = "1E5954CB15B6F24A00A03429"
-                     BuildableName = "libAppInsights.a"
-                     BlueprintName = "AppInsights"
-                     ReferencedContainer = "container:AppInsights.xcodeproj">
-                  </BuildableReference>
-               </EnvironmentBuildable>
-            </ActionContent>
-         </ExecutionAction>
-      </PostActions>
+      buildConfiguration = "Debug">
       <Testables>
          <TestableReference
             skipped = "NO">
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "1E5A458F16F0DFC200B55C04"
-               BuildableName = "AppInsightsTests.xctest"
-               BlueprintName = "AppInsights Tests"
-               ReferencedContainer = "container:AppInsights.xcodeproj">
+               BuildableName = "ApplicationInsightsTests.xctest"
+               BlueprintName = "ApplicationInsights Tests"
+               ReferencedContainer = "container:ApplicationInsights.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -61,9 +42,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "1E5954CB15B6F24A00A03429"
-            BuildableName = "libAppInsights.a"
-            BlueprintName = "AppInsights"
-            ReferencedContainer = "container:AppInsights.xcodeproj">
+            BuildableName = "libApplicationInsights.a"
+            BlueprintName = "ApplicationInsights"
+            ReferencedContainer = "container:ApplicationInsights.xcodeproj">
          </BuildableReference>
       </MacroExpansion>
    </TestAction>
@@ -80,9 +61,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "1E5954CB15B6F24A00A03429"
-            BuildableName = "libAppInsights.a"
-            BlueprintName = "AppInsights"
-            ReferencedContainer = "container:AppInsights.xcodeproj">
+            BuildableName = "libApplicationInsights.a"
+            BlueprintName = "ApplicationInsights"
+            ReferencedContainer = "container:ApplicationInsights.xcodeproj">
          </BuildableReference>
       </MacroExpansion>
       <AdditionalOptions>
@@ -98,9 +79,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "1E5954CB15B6F24A00A03429"
-            BuildableName = "libAppInsights.a"
-            BlueprintName = "AppInsights"
-            ReferencedContainer = "container:AppInsights.xcodeproj">
+            BuildableName = "libApplicationInsights.a"
+            BlueprintName = "ApplicationInsights"
+            ReferencedContainer = "container:ApplicationInsights.xcodeproj">
          </BuildableReference>
       </MacroExpansion>
    </ProfileAction>
