+++ Support/module.modulemap
@@ -1,5 +1,5 @@
-framework module AppInsights {
-  umbrella header "AppInsights.h"
+framework module ApplicationInsights {
+  umbrella header "ApplicationInsights.h"
   
   export *
   module * { export * }
