+++ docs/Guide-Installation-Setup-template.md
@@ -39,14 +39,14 @@ The SDK runs on devices with iOS 6.0 or higher.
 
 1. Download the latest [Application Insights for iOS](https://github.com/Microsoft/AppInsights-iOS/releases) framework.
 
-2. Unzip the file. A new folder `AppInsights` is created.
+2. Unzip the file. A new folder `ApplicationInsights` is created.
 
 3. Move the folder into your project directory. We usually put 3rd-party code into a subdirectory named `Vendor`, so we move the directory into it.
 
 <a id="xcode"></a> 
 ## Set up Xcode
 
-1. Drag & drop `AppInsights.framework` from your project directory to your Xcode project.
+1. Drag & drop `ApplicationInsights.framework` from your project directory to your Xcode project.
 2. Similar to above, our projects have a group `Vendor`, so we drop it there.
 3. Select `Create groups for any added folders` and set the checkmark for your target. Then click `Finish`.
 4. Select your project in the `Project Navigator` (⌘+1).
@@ -70,23 +70,23 @@ The SDK runs on devices with iOS 6.0 or higher.
 3. Add the following line at the top of the file below your own #import statements:
 
 	```objectivec
-	#import <AppInsights/AppInsights.h>
+	#import <ApplicationInsights/ApplicationInsights.h>
 	```
 4. Search for the method `application:didFinishLaunchingWithOptions:`
-5. Add the following lines to setup and start the AppInsights SDK:
+5. Add the following lines to setup and start the Application Insights SDK:
 
 	```objectivec
-	[[MSAIAppInsights sharedInstance] setup];
+	[[MSAIApplicationInsights sharedInstance] setup];
 	// Do some additional configuration if needed here
 	...
-	[[MSAIAppInsights sharedInstance] start];
+	[[MSAIApplicationInsights sharedInstance] start];
 	```
 
 	You can also use the following shortcut:
 
 	```objectivec
-	[MSAIAppInsights setup];
-	[MSAIAppInsights start];
+	[MSAIApplicationInsights setup];
+	[MSAIApplicationInsights start];
 	```
 
 6. Send some data to the server:
@@ -121,25 +121,25 @@ The SDK runs on devices with iOS 6.0 or higher.
 3. Add the following line at the top of the file below your own #import statements:
 	
 	```swift	
-	#import AppInsights
+	#import ApplicationInsights
 	```
 4. Search for the method 
 	
 	```swift	
 	application(application: UIApplication, didFinishLaunchingWithOptions launchOptions:[NSObject: AnyObject]?) -> Bool`
 	```
-5. Add the following lines to setup and start the AppInsights SDK:
+5. Add the following lines to setup and start the Application Insights SDK:
 	
 	```swift
-	MSAIAppInsights.sharedInstance().setup();
-   MSAIAppInsights.sharedInstance().start();
+	MSAIApplicationInsights.sharedInstance().setup();
+   MSAIApplicationInsights.sharedInstance().start();
 	```
 	
 	You can also use the following shortcut:
 
 	```swift
-	MSAIAppInsights.setup();
-   MSAIAppInsights.start();
+	MSAIApplicationInsights.setup();
+   MSAIApplicationInsights.start();
 	```
 5. Send some data to the server:
 	
@@ -175,9 +175,9 @@ By default the following server URL is used to work with the [Azure portal](http
 To change the URL, setup Application Insights like this:
 
 ```objectivec
-	[[MSAIAppInsights sharedInstance] setup];
-	[[MSAIppInsights sharedInstance] setServerURL:{your server url}];
-	[[MSAIAppInsights sharedInstance] start];
+	[[MSAIApplicationInsights sharedInstance] setup];
+	[[MSAIApplicationInsights sharedInstance] setServerURL:{your server url}];
+	[[MSAIApplicationInsights sharedInstance] start];
 ```
 	
 
@@ -191,17 +191,17 @@ The following points need to be considered to use the Application Insights SDK w
 
 	```objectivec	
 	@interface TodayViewController () <NCWidgetProviding>
-	@property (nonatomic, assign) BOOL didSetupAppInsightsSDK;
+	@property (nonatomic, assign) BOOL didSetupApplicationInsightsSDK;
 	@end
 
 	@implementation TodayViewController
 
 	- (void)viewDidLoad {
 		[super viewDidLoad];
-		if (!self.didSetupAppInsightsSDK) {
-			[MSAIAppInsights setup];
-			[MSAIAppInsights start];
-          self.didSetupAppInsightsSDK = YES;
+		if (!self.didSetupApplicationInsightsSDK) {
+			[MSAIApplicationInsights setup];
+			[MSAIApplicationInsights start];
+          self.didSetupApplicationInsightsSDK = YES;
        }
     }
     ```
@@ -221,17 +221,17 @@ Instead of manually adding the missing frameworks, you can also use our bundled 
 
 4. Expand `Configurations`.
 
-5. Select `AppInsights.xcconfig` for all your configurations (if you don't already use a `.xcconfig` file)
+5. Select `ApplicationInsights.xcconfig` for all your configurations (if you don't already use a `.xcconfig` file)
 
 	**Note:** You can also add the required frameworks manually to your targets `Build Phases` and continue with step `7.` instead.
 
 6. If you are already using a `.xcconfig` file, simply add the following line to it
 
-	`#include "../Vendor/AppInsights/Support/AppInsights.xcconfig"`
+	`#include "../Vendor/ApplicationInsights/Support/ApplicationInsights.xcconfig"`
 
 	(Adjust the path depending where the `Project.xcconfig` file is located related to the Xcode project package)
 
-	**Important note:** Check if you overwrite any of the build settings and add a missing `$(inherited)` entry on the projects build settings level, so the `AppInsights.xcconfig` settings will be passed through successfully.
+	**Important note:** Check if you overwrite any of the build settings and add a missing `$(inherited)` entry on the projects build settings level, so the `ApplicationInsights.xcconfig` settings will be passed through successfully.
 
 7. If you are getting build warnings, then the `.xcconfig` setting wasn't included successfully or its settings in `Other Linker Flags` get ignored because `$(inherited)` is missing on project or target level. Either add `$(inherited)` or link the following frameworks manually in `Link Binary With Libraries` under `Build Phases`:
 	- `UIKit`
@@ -258,4 +258,4 @@ pod "ApplicationInsights", '1.0-beta.1'
 <a id="contact"></a>
 ## Contact
 
-If you have further questions or are running into trouble that cannot be resolved by any of the steps here, feel free to contact us at [AppInsights-iOS@microsoft.com](mailto:appinsights-ios@microsoft.com)
+If you have further questions or are running into trouble that cannot be resolved by any of the steps here, feel free to contact us at [AppInsights-iOS@microsoft.com](mailto:AppInsights-ios@microsoft.com)
